Move the fetch store into its own module
InMemoryIndex is the one object the fetcher and the provider both hold, and
it is stdlib-only: no network, no filesystem, no index client at runtime.
fetch.py is where the asyncio coordinator and the HTTP stack behind it live,
so the store sat behind an import of both.

store.py holds four definitions. _Pending moves byte for byte. The two
pending-key helpers move byte for byte apart from their names, _metadata_key
to metadata_pending_key and _range_key to range_pending_key, because a
private name cannot be imported across the module boundary fetch.py now
crosses. InMemoryIndex is 542 lines in both trees, and five of them differ:
the two calls to metadata_pending_key and the three to range_pending_key.
The module docstring, the imports and the three-name __all__ are the file's
own.

fetch.py keeps re-exporting InMemoryIndex, so no import site had to move.
metadata_resolver.py's type-only import moves anyway, with the class's name
in one docstring cross-reference, to name the module the class now lives in.
